requirements.txt
channels==2.1.2
channels-redis
django
websocket-client
djangorestframework
channels-rabbitmq==0.0.11
